+++ translator/test_infra/sol/operators/bitwise/ignore_simple.sol
@@ -5,8 +5,8 @@
 pragma solidity ^0.8.10;
 
 contract Simple {
-    // # not_uint (20) 65515
-    function not_uint(uint a) public pure returns (uint) {
+    // # not_uint128 (20) 65515
+    function not_uint(uint128 a) public pure returns (uint128) {
         return ~a;
     }
 }
